+++ app/commands/characters_context/cosmere/create_command.rb
@@ -14,6 +14,7 @@ module CharactersContext
           required(:user).filled(type?: User)
           required(:name).filled(:string, max_size?: 50)
           required(:ancestry).filled(:string)
+          required(:setting).filled(:string)
           required(:cultures).filled(:array, max_size?: 2).each(:string)
           optional(:path).filled(:string)
           optional(:skip_guide).filled(:bool)
@@ -23,7 +24,7 @@ module CharactersContext
       private
 
       def do_prepare(input)
-        input[:data] = build_fresh_character(input.slice(:ancestry, :cultures, :path, :skip_guide).symbolize_keys)
+        input[:data] = build_fresh_character(input.slice(:ancestry, :cultures, :path, :skip_guide, :setting).symbolize_keys)
         input[:initial_talents] = input[:data].delete(:initial_talents)
       end