+++ db/migrate/20260903042305_backfill_cosmere_settings.rb
@@ -1,0 +1,8 @@
+class BackfillCosmereSettings < ActiveRecord::Migration[8.1]
+  def change
+    Cosmere::Character.find_each do |character|
+      character.data.setting = 'roshar'
+      character.save
+    end
+  end
+end